+++ ds-license/CHANGELOG.md
@@ -16,7 +16,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - new property in ds-license-behaviour.yaml for cutoff year for TV. All TV before this year will bypass holdback rules.
 
 ### Changed
-- Allow all TV before 1973 (year defined in property file). ProductionCodeAllowed will be true and holdbackExpired will be start time for program.
+- Allow all TV before 1974(not included). The cutoff year can be configured in the yaml-file. ProductionCodeAllowed will be true and holdbackExpired will be start time for program.
 - Formatted `openapi` file.
 - Moved OpenAPI description from `pom.xml` into `openapi` file.
 - Added `name` in `pom.xml`.